Test The Commit

+++ Backend/Utilities/Cloudinary.js
@@ -1,10 +1,10 @@
-const cloudinary = require(cloudinary).v2
+const cloudinary = require("cloudinary").v2;
 const fs = require("fs")
 
-cloudinary.config({ 
-  cloud_name: process.env.CLOUDINARY_CLOUD_NAME, 
-  api_key: process.env.CLOUDINARY_API_KEY, 
-  api_secret: process.env.CLOUDINARY_API_SECRET 
+cloudinary.config({
+    cloud_name: process.env.CLOUDINARY_CLOUD_NAME,
+    api_key: process.env.CLOUDINARY_API_KEY,
+    api_secret: process.env.CLOUDINARY_API_SECRET
 });
 
 const uploadOnCloudinary = async (localFilePath) => {
@@ -16,8 +16,14 @@ const uploadOnCloudinary = async (localFilePath) => {
         })
         // file has been uploaded successfull
         //console.log("file is uploaded on cloudinary ", response.url);
-        fs.unlinkSync(localFilePath)
+        if (fs.existsSync(localFilePath)) {
+            fs.unlinkSync(localFilePath);
+        }
+
+        console.log(response);
+
         return response;
+
 
     } catch (error) {
         fs.unlinkSync(localFilePath) // remove the locally saved temporary file as the upload operation got failed
@@ -27,4 +33,4 @@ const uploadOnCloudinary = async (localFilePath) => {
 
 
 
-module.exports =  {uploadOnCloudinary}
+module.exports = { uploadOnCloudinary }
